ci: remove regular build from CI
Something in the underlying image changed, and the CMake build does not
find rustc compiler anymore (despite rustc still being available). The
effort to reproduce the setup locally and find the issue, as well as
similar maintenance in the future, is not worth the considerable amount
of time. Thus, we remove the regular build from the CI and keep only the
reproducable Nix-based builds.